Added Commenting

+++ app/src/main/java/com/example/s1636469/coinz/ProfileActivity.java
@@ -1,6 +1,12 @@
+/*
+ *  ProfileActivity
+ *
+ *  Displays player information and allows users to add and remove friends
+ *
+ */
+
 package com.example.s1636469.coinz;
 
-import android.content.Intent;
 import android.graphics.Bitmap;
 import android.os.Bundle;
 import android.app.Activity;
@@ -18,40 +24,61 @@ import com.google.android.gms.tasks.OnFailureListener;
 import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseUser;
 import com.google.firebase.firestore.DocumentReference;
 import com.google.firebase.firestore.DocumentSnapshot;
 import com.google.firebase.firestore.FirebaseFirestore;
-import com.google.firebase.firestore.SetOptions;
 import com.mikhaellopez.circularimageview.CircularImageView;
 
-import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.Map;
+import java.util.Set;
 
 public class ProfileActivity extends Activity {
     protected String TAG = "Profile";
     private String friend_id;
+    private FloatingActionButton friend_button;
+
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_profile);
+        friend_button =  findViewById(R.id.profile_friend_button);
+        friend_button.setEnabled(true);
         fillInValues();
 
-        Intent i = getIntent();
-        boolean is_friend = i.getBooleanExtra("already_friends",false);
         friend_id = FriendListFragment.pass_to_profile.getString("id");
+        checkIfFriends();
+
+    }
+    /*
+     *  checkIfFriends
+     *
+     *  check the users's friends collection, if they aer friends then return true else false
+     */
+    private void checkIfFriends() {
 
         String user_id = FirebaseAuth.getInstance().getUid();
+        if (user_id == null) {
+            friend_button.setEnabled(false);
+            Toast.makeText(getApplicationContext(), "Cannot get friend data right now, so you " +
+                    "cannot manage thier friend status", Toast.LENGTH_SHORT).show();
+            return;
+        }
+
         FirebaseFirestore database = FirebaseFirestore.getInstance();
         DocumentReference f_ref = database.collection("users")
-                .document("user_id")
+                .document(user_id)
                 .collection("friends")
                 .document(friend_id);
+
+        // Get friend collection
         f_ref.get().addOnCompleteListener(new OnCompleteListener<DocumentSnapshot>() {
             @Override
             public void onComplete(@NonNull Task<DocumentSnapshot> task) {
                 if (task.isSuccessful()) {
-                    if (task.getResult().exists()) {
+                    // if the friend collection has the user
+                    if (task.getResult() != null && task.getResult().exists()) {
                         setUpListeners(true);
                     } else {
                         setUpListeners(false);
@@ -62,9 +89,15 @@ public class ProfileActivity extends Activity {
             }
         });
     }
-
+    /*
+     *  fillInValues
+     *
+     *  update UI with data from Bundle
+     */
     protected void fillInValues() {
         if (FriendListFragment.pass_to_profile != null ){
+
+            // Get all data
             String name = (String) FriendListFragment.pass_to_profile.get("name");
             Bitmap img = (Bitmap) FriendListFragment.pass_to_profile.get("img");
             HashMap<String, Double> curs;
@@ -75,17 +108,24 @@ public class ProfileActivity extends Activity {
                 return;
             }
 
-
-            CircularImageView profile_img = (CircularImageView) findViewById(R.id.profile_img_preview);
+            // Display Data
+            CircularImageView profile_img =  findViewById(R.id.profile_img_preview);
             profile_img.setImageBitmap(img);
 
-            TextView name_view = (TextView) findViewById(R.id.profile_name);
+            TextView name_view =  findViewById(R.id.profile_name);
             name_view.setText(name);
 
-            LinearLayout ll = (LinearLayout) findViewById(R.id.currency_amount_filler);
+            // fill in all currencies
+            LinearLayout ll =  findViewById(R.id.currency_amount_filler);
+            if (curs == null) {
+                Log.d(TAG, "Currency is null!");
+                Toast.makeText(getApplicationContext(), "Cannot find currency values for player",Toast.LENGTH_SHORT).show();
+                return;
+            }
+
             for (String k : curs.keySet()) {
                 TextView tv = new TextView(this);
-                Log.d(TAG,((Double)curs.get(k)).toString());
+                Log.d(TAG,(curs.get(k)).toString());
                 tv.setText(k + ": " + curs.get(k).toString());
                 tv.setTextAlignment(View.TEXT_ALIGNMENT_CENTER);
                 tv.setTextSize(20f);
@@ -93,9 +133,13 @@ public class ProfileActivity extends Activity {
             }
         }
     }
-
+    /*
+     *  setUpListeners
+     *
+     *  sets up listeners for all of the buttons in the activity
+     */
     private void setUpListeners(boolean is_friend) {
-        ImageView close_icon = (ImageView) findViewById(R.id.close_profile);
+        ImageView close_icon = findViewById(R.id.close_profile);
         close_icon.setOnClickListener(new View.OnClickListener() {
             @Override
             public void onClick(View v) {
@@ -105,54 +149,32 @@ public class ProfileActivity extends Activity {
         updateButton(is_friend);
     }
 
+    /*
+     *  addFriend
+     *
+     *  Add the player shown in profile as a friend
+     *
+     */
     private void addFriend() {
+        // Get reference to the database
         FirebaseFirestore database = FirebaseFirestore.getInstance();
+
         FirebaseAuth auth = FirebaseAuth.getInstance();
+        FirebaseUser user = auth.getCurrentUser();
+        if (user == null) {
+            Log.d(TAG, "User was null");
+            Toast.makeText(getApplicationContext(), "Cannot add as friend",Toast.LENGTH_SHORT).show();
+            return;
+        }
         String id = auth.getCurrentUser().getUid();
+
         DocumentReference user_ref = database.collection("users").document(id);
-        Log.d(TAG, "friend name %s".format(friend_id));
+        // Get user
         DocumentReference friend_ref = database.collection("users").document(friend_id);
         friend_ref.get().addOnSuccessListener(new OnSuccessListener<DocumentSnapshot>() {
             @Override
             public void onSuccess(DocumentSnapshot documentSnapshot) {
-                Map<String, Object> data = (Map<String, Object>) documentSnapshot.getData();
-                if (data != null) {
-                    String profile_url = (String) data.get("profile_url");
-                    String profile_name = (String) data.get("name");
-                    String id = documentSnapshot.getId();
-
-
-                    HashMap<String, Object> new_friend = new HashMap<>();
-                    new_friend.put("name",profile_name);
-                    new_friend.put("profile_url", profile_url);
-
-                    try {
-                        new_friend.put("GOLD",(Double) data.get("GOLD"));
-                    } catch (ClassCastException e) {
-                        new_friend.put("GOLD",((Long) data.get("GOLD")).doubleValue());
-                    }
-
-                    user_ref.collection("friends").document(id).set(new_friend)
-                            .addOnSuccessListener(new OnSuccessListener<Void>() {
-                        @Override
-                        public void onSuccess(Void aVoid) {
-                            updateButton(true);
-                            Toast.makeText(getApplicationContext(), "Added " + profile_name +
-                                "as friend",Toast.LENGTH_SHORT).show();
-                        }
-                    }).addOnFailureListener(new OnFailureListener() {
-                        @Override
-                        public void onFailure(@NonNull Exception e) {
-                            Toast.makeText(getApplicationContext(), "Cannot add friend",
-                                    Toast.LENGTH_SHORT).show();
-                            Log.w(TAG,"Error thrown when adding friend",e);
-                        }
-                    });
-                } else {
-                    Log.w(TAG, "friend data is null!");
-                    Toast.makeText(getApplicationContext(),"Cannot add friend!",Toast.LENGTH_SHORT).show();
-                }
-
+                addFriendFirebaseUpdate(documentSnapshot, user_ref);
             }
         }).addOnFailureListener(new OnFailureListener() {
             @Override
@@ -162,13 +184,80 @@ public class ProfileActivity extends Activity {
             }
         });
     }
+    /*
+     *  addFriendFirebaseUpdate
+     *
+     *  Update FireStore with new friend
+     */
+    private void addFriendFirebaseUpdate(DocumentSnapshot documentSnapshot, DocumentReference user_ref) {
+        Map<String, Object> data = documentSnapshot.getData();
+        if (data != null) {
 
+            // Get the nessesary data about the player
+            String profile_url = (String) data.get("profile_url");
+            String profile_name = (String) data.get("name");
+            String id = documentSnapshot.getId();
+
+            // Get data to push
+            HashMap<String, Object> new_friend = new HashMap<>();
+            new_friend.put("name",profile_name);
+            new_friend.put("profile_url", profile_url);
+
+            try {
+                new_friend.put("GOLD",(Double) data.get("GOLD"));
+            } catch (ClassCastException e) {
+                new_friend.put("GOLD",((Long) data.get("GOLD")).doubleValue());
+            }
+
+            // Push the data and notify user
+            user_ref.collection("friends").document(id).set(new_friend)
+                    .addOnSuccessListener(new OnSuccessListener<Void>() {
+                        @Override
+                        public void onSuccess(Void aVoid) {
+                            updateButton(true);
+                            Toast.makeText(getApplicationContext(), "Added " + profile_name +
+                                    "as friend",Toast.LENGTH_SHORT).show();
+                        }
+                    }).addOnFailureListener(new OnFailureListener() {
+                @Override
+                public void onFailure(@NonNull Exception e) {
+                    Toast.makeText(getApplicationContext(), "Cannot add friend",
+                            Toast.LENGTH_SHORT).show();
+                    Log.w(TAG,"Error thrown when adding friend",e);
+                }
+            });
+
+        } else {
+            Log.w(TAG, "friend data is null!");
+            Toast.makeText(getApplicationContext(),"Cannot add friend!",Toast.LENGTH_SHORT).show();
+        }
+    }
+
+    /*
+     *  removeFriend
+     *
+     *  removes friend from user's list
+     */
     public void removeFriend() {
+
+        // Get reference to firestore
         FirebaseFirestore database = FirebaseFirestore.getInstance();
+
+
         FirebaseAuth auth = FirebaseAuth.getInstance();
+        FirebaseUser user = auth.getCurrentUser();
+        if (user == null) {
+            Log.d(TAG, "User was null");
+            Toast.makeText(getApplicationContext(), "Cannot remove friend",Toast.LENGTH_SHORT).show();
+            return;
+        }
         String id = auth.getCurrentUser().getUid();
+
+        // Get reference to remvoe
         DocumentReference remove_ref = database.collection("users").document(id)
                 .collection("friends").document(friend_id);
+
+        // Delete the document from the user's friends collection
         remove_ref.delete().addOnSuccessListener(new OnSuccessListener<Void>() {
             @Override
             public void onSuccess(Void aVoid) {
@@ -184,12 +273,15 @@ public class ProfileActivity extends Activity {
             }
         });
     }
-
+    /*
+     *  updateButton
+     *
+     *  change the add/remove button's functionality
+     *
+     */
     private void updateButton(boolean is_friend) {
-        Log.d(TAG,"update button");
-        Log.d(TAG, "userid " + friend_id);
-        FloatingActionButton friend_button = (FloatingActionButton) findViewById(R.id.profile_friend_button);
         if (is_friend) {
+            // Change image on button and onclick listener
             friend_button.setImageResource(R.drawable.remove_white);
             friend_button.setOnClickListener(new View.OnClickListener() {
                 @Override
@@ -200,6 +292,8 @@ public class ProfileActivity extends Activity {
                 }
             });
         } else {
+            // Change image on button and onclick listener
+
             friend_button.setImageResource(R.drawable.add_white);
             friend_button.setOnClickListener(new View.OnClickListener() {
                 @Override
@@ -210,6 +304,7 @@ public class ProfileActivity extends Activity {
                 }
             });
         }
+        // enable button
         friend_button.setEnabled(true);
     }
 }
